Removed code dup, moved load weights function to _base.py
