+++ pyserini/encode/_unicoil.py
@@ -21,9 +21,9 @@ import torch
 if torch.cuda.is_available():
     from torch.cuda.amp import autocast
 from transformers import BertConfig, BertModel, BertTokenizer, PreTrainedModel
-from transformers.utils import cached_file
 
 from pyserini.encode import DocumentEncoder, QueryEncoder
+from pyserini.encode._base import load_head_weights
 from packaging.version import Version
 from transformers import __version__ as transformers_version
 
@@ -79,13 +79,6 @@ class UniCoilEncoder(PreTrainedModel):
         tok_weights = self.tok_proj(sequence_output)
         tok_weights = torch.relu(tok_weights)
         return tok_weights
-
-def load_head_weights(model, model_name, weight_map):
-    weights_path = cached_file(model_name, 'pytorch_model.bin')
-    state_dict = torch.load(weights_path, map_location='cpu', weights_only=True)
-    for module_name, keys in weight_map.items():
-        module = getattr(model, module_name)
-        module.load_state_dict({k: state_dict[v] for k, v in keys.items()})
 
 class UniCoilDocumentEncoder(DocumentEncoder):
     def __init__(self, model_name, tokenizer_name=None, device='cuda:0'):
